finish the calculation and writing parts (need test)
Now the calculation is divided into 3 parts: first we
calculate the all the coord_pairs that we need translate (include
triple vertices cases and also considering the filter according to
the input argument).
Then we calculate the background peptide and kmer.
Finally, we caculate the junction kmer for each simple_metadata and
write them at once.

Some minor changes are:
We create a filepointera namedtuple and pass it onto the immuno_model.py.

Need further test as it is not right in the concatenate part for the
simple check. Also the naming of junction result is not so clear --
seem to be abusing the `.` -- need a better symbol.