Remove the assertion in checking if somatic mutations are consistency. (#31)
Originally, if there is an error, we will skip the current gene and go to
the next one. After discussion, we think this error is not important.
Sometimes there might be some inconsistency between maf file and
genome for the reference gene. It is however not a problem for our
case (will not bring new kmer or miss kmer). In this way, there is no
need to skip the gene.