Improve the logging module Now the log file is saved in output_dir/error.log. All the print functions are already converted into logging. Currently there are three levels used. logging.info, logging.warning and logging.error.
-Info record the time and memory cost of different stages. Also will
show the summary statistics for each complete round, that is finishing
traversing all splicegraphs for one samples.

-Warning usually happens when some infomation is missing. Eg, count file or
mutation file. The reason might be that they do not exist and the users are
recommended to have a double check. So when the information is missing, we will
still proceed without that information. In this case, the computation might be
redundant.

-Error usually happens when some unexpected things happen and we can not proceed.
I have already pointed out some places where error might exist with assert sentence.
In other places, there will be try-exception block to capture the error information and
write into the error.log file.

-Error